requirements.txt
streamlit>=1.28.0,<2.0.0
pandas>=2.0.0,<2.3.0
spacy>=3.6.0,<3.9.0
anthropic>=0.7.0
google-generativeai>=0.3.0
requests>=2.31.0

# Core scientific packages
# Note: For Python 3.12+, use numpy>=1.26.0 and scipy>=1.11.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<1.14.0

# Build tools (important for Python 3.12+)
setuptools>=65.0.0
wheel>=0.38.0

# Additional dependencies that might be needed
# If you encounter issues, uncomment these:
# scikit-learn>=1.3.0,<1.5.0
# matplotlib>=3.7.0,<3.10.0